Create 500mbheightvort.gs
This script plots 500 hPa heights/absolute vorticity from the GFS in decameters and inverse seconds, respectively, over a polar stereographic projection of the contiguous U.S. and some of Canada/Mexico. It then saves the output to a .png file named 500-hPa-Height-Vorticity in the current directory. To run this script,you will need to have xcbar.gs and colormaps.gs installed. 

Made with OpenGrads 2.1.0 Linux 32-bit.